runit-services.bbclass busybox: Standardize selecting busybox-runit-init as init system
Rather than using a combination of settings to get busybox-runit-init as
the init system, only make selection of busybox-runit-init as init depend
on the VIRTUAL_RUNTIME_init_manager setting.  This incidentally leaves
sysvinit as the default init DISTRO_FEATURES setting, which means that
the various initscripts get properly installed and can be used by
runit-init, when necessary, without modifying all the conditionalized
on systemvinit settings.

Signed-off-by: Daniel Dickinson <cshored@cshore.thecshore.com>